requirements.txt
fastapi
uvicorn
pydantic
pdfminer.six
python-docx
spacy
textstat
sentence-transformers
regex
numpy
